feat(athena-driver): Migrate to TypeScript
