+++ packages/cubejs-athena-driver/index.js
@@ -1,0 +1,15 @@
+const fromExports = require('./dist/src');
+const { AthenaDriver } = require('./dist/src/AthenaDriver');
+
+/**
+ * After 5 years working with TypeScript, now I know
+ * that commonjs and nodejs require is not compatibility with using export default
+ */
+const toExport = AthenaDriver;
+
+// eslint-disable-next-line no-restricted-syntax
+for (const [key, module] of Object.entries(fromExports)) {
+  toExport[key] = module;
+}
+
+module.exports = toExport;
