+++ packages/cubejs-athena-driver/package.json
@@ -2,7 +2,7 @@
   "name": "@cubejs-backend/athena-driver",
   "description": "Cube.js Athena database driver",
   "author": "Cube Dev, Inc.",
-  "version": "0.28.4",
+  "version": "0.28.0",
   "repository": {
     "type": "git",
     "url": "https://github.com/cube-js/cube.js.git",
@@ -11,19 +11,28 @@
   "engines": {
     "node": ">=10.8.0"
   },
+  "files": [
+    "dist/src",
+    "index.js"
+  ],
+  "main": "index.js",
+  "typings": "dist/src/index.d.ts",
   "scripts": {
-    "lint": "eslint **/*.js"
+    "build": "rm -rf dist && npm run tsc",
+    "tsc": "tsc",
+    "watch": "tsc -w",
+    "lint": "eslint src/* --ext .ts",
+    "lint:fix": "eslint --fix src/* --ext .ts"
   },
-  "main": "driver/AthenaDriver.js",
-  "types": "driver/index.d.ts",
   "dependencies": {
     "@aws-sdk/client-athena": "^3.22.0",
-    "@cubejs-backend/query-orchestrator": "^0.28.3",
-    "@cubejs-backend/shared": "^0.28.2",
+    "@cubejs-backend/query-orchestrator": "^0.28.0",
+    "@cubejs-backend/shared": "^0.28.0",
     "sqlstring": "^2.3.1"
   },
   "devDependencies": {
-    "@cubejs-backend/linter": "^0.28.0"
+    "@cubejs-backend/linter": "^0.28.0",
+    "typescript": "~4.1.6"
   },
   "eslintConfig": {
     "extends": "../cubejs-linter"
