+++ packages/cubejs-athena-driver/src/index.ts
@@ -1,0 +1,5 @@
+import { AthenaDriver } from './AthenaDriver';
+
+export * from './AthenaDriver';
+
+export default AthenaDriver;
