+++ tsconfig.json
@@ -35,6 +35,9 @@
       "path": "packages/cubejs-postgres-driver"
     },
     {
+      "path": "packages/cubejs-athena-driver"
+    },
+    {
       "path": "packages/cubejs-mongobi-driver"
     },
     {
